+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # WIP/Unreleased
 
+- Fix issue where stdout from a kubetools dev exception would not be formatted properly
+- Add "dev backends" support for future work (alternatives to `docker-compose`)
+
 # v9.1.1
 - Add ability to define number of retries on readinessProbe